TASK-226: fix set cost when we start to wait for smart deposit and don't change it until 'wait_for' time

+++ dynamis/apps/payments/api/v1/serializers.py
@@ -10,7 +10,7 @@ from dynamis.utils.math import approximately_equal
 class SmartDepositShortSerializer(serializers.ModelSerializer):
     status = serializers.IntegerField(source='state')
     cost_in_eth = serializers.FloatField(source='cost')
-    cost_in_wei = serializers.FloatField(source='cost_wei')
+    cost_in_wei = serializers.IntegerField(source='cost_wei')
     cost_in_dollars = serializers.FloatField(source='cost_dollar')
     address_to_send = SerializerMethodField()
